+++ AgendamentoObject/GerarCodigo.java
@@ -1,0 +1,5 @@
+package agendamento0bject;
+
+interface GeradorCodigo {
+    int gerarCodigo();
+}
